+++ test/options.js
@@ -1,0 +1,4 @@
+module.exports = {
+  host :  'http://surge-server.cloudapp.net:8080',
+  debug:  false
+}